translation-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
google-generativeai==0.3.1
httpx==0.25.0
python-jose==3.3.0
pydantic==2.4.2
